--- lib/src/ui/components/data_widgets/roc_dropdown_menu_item.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Roc's custom top application bar.
-class RocDropdownMenuItem<T> extends DropdownMenuItem {
-  RocDropdownMenuItem(
-    BuildContext context,
-    MapEntry<T, String> entry,
-  ) : super(
-          value: entry.key,
-          alignment: Alignment.center,
-          child: Text(
-            entry.value,
-            style: Theme.of(context).textTheme.bodyMedium,
-          ),
-        );
-}
