--- lib/src/ui/components/view_widgets/roc_chip_view.dart
@@ -1,35 +1,0 @@
-import 'package:darq/darq.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:logger/logger.dart';
-
-import '../data_widgets/roc_data_chip.dart';
-import '../data_widgets/roc_warning_chip.dart';
-
-/// Roc's custom chip column view widget.
-class RocChipView extends StatelessWidget {
-  final List<String> _entries;
-  final Logger _logger;
-
-  RocChipView(List<String> entries, Logger logger)
-      : _entries = entries,
-        _logger = logger;
-
-  List<Widget> formChipsList(BuildContext context) {
-    if (_entries.isEmpty) {
-      return List<Widget>.from(
-          [RocWarningChip(AppLocalizations.of(context)!.noIPsShortWarning)]);
-    }
-
-    return _entries
-        .select<Widget>((entry, i) => RocDataChip(entry, _logger))
-        .toList();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: formChipsList(context),
-    );
-  }
-}
