+++ test/ui/pages/about_page_test.dart
@@ -2,10 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:roc_droid/src/model.dart';
 import 'package:roc_droid/src/ui.dart';
-import 'package:roc_droid/src/ui/components/input_widgets/roc_about_page_button.dart';
-import 'package:roc_droid/src/ui/components/scaffold_widgets/roc_side_pane_app_bar.dart';
 import 'package:roc_droid/src/ui/components/view_widgets/roc_scroll_view.dart';
-import 'package:roc_droid/src/ui/fragments/about_page.dart';
+import 'package:roc_droid/src/ui/pages/about_page.dart';
 import 'package:roc_droid/src/ui/utils/roc_keys.dart';
 
 // Side pane class widget tests.
@@ -19,7 +17,7 @@ void main() {
     // Find required widgets
     final sidePane = find.byType(AboutPage);
     final scaffold = find.byType(Scaffold);
-    final appBar = find.byType(RocSidePaneAppBar);
+    final appBar = find.byType(RocAboutPageAppBar);
     final scrollView = find.byType(RocScrollView);
     final aboutPageButtons = find.byType(RocAboutPageButton);
