+++ test/ui/pages/sender_page_test.dart
@@ -2,13 +2,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:roc_droid/src/model.dart';
 import 'package:roc_droid/src/model/entities/capture_source_type.dart';
 import 'package:roc_droid/src/ui.dart';
-import 'package:roc_droid/src/ui/components/data_widgets/roc_port_chip.dart';
+import 'package:roc_droid/src/ui/components/data_widgets/roc_chips.dart';
 import 'package:roc_droid/src/ui/components/data_widgets/roc_text_row.dart';
 import 'package:roc_droid/src/ui/components/input_widgets/roc_dropdown_button.dart';
-import 'package:roc_droid/src/ui/components/input_widgets/roc_number_input.dart';
 import 'package:roc_droid/src/ui/components/input_widgets/roc_stateful_button.dart';
 import 'package:roc_droid/src/ui/components/view_widgets/roc_page_view.dart';
-import 'package:roc_droid/src/ui/sender_page.dart';
+import 'package:roc_droid/src/ui/pages/sender_page.dart';
 import 'package:roc_droid/src/ui/utils/roc_keys.dart';
 
 // Sender page class widget tests.
